requirements.txt
-i https://pypi.org/simple
amqp==2.3.2
billiard==3.5.0.4
boto3==1.9.40
botocore==1.12.50
celery[solar]==4.2.1
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
dj-email-url==0.1.0
django-celery-beat==1.3.0
django-configurations==2.1
django-extensions==2.1.3
django-ical==1.5
django-piston==0.2.3
django-storages==1.7.1
django-timezone-field==3.0
django==2.0.9
djangorestframework-xml==1.3.0
djangorestframework==3.9.0
docutils==0.14
envdir==1.0.1
ephem==3.7.6.0
gunicorn==19.9.0
icalendar==4.0.3
idna==2.7
jmespath==0.9.3
kombu==4.2.1
pillow==5.3.0
psycopg2==2.7.5
pylast==2.4.0
python-crontab==2.3.5
python-dateutil==2.7.5 ; python_version >= '2.7'
pytz==2018.7
raven==6.9.0
redis==2.10.6
requests==2.20.1
s3transfer==0.1.13
six==1.11.0
urllib3==1.24.1 ; python_version >= '3.4'
vine==1.1.4
